requirements.txt
flask
yt-dlp
playwright
requests
gunicorn
flask-sqlalchemy
python-dotenv